// src/lib/api/generated/model/v1Group.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Aisphere IAM API
 * OpenAPI spec version: v1
 */

export interface V1Group {
  displayName?: string;
  externalId?: string;
  id?: string;
  name?: string;
  orgId?: string;
  parentId?: string;
  path?: string;
  type?: string;
  users?: string[];
}